Add test components
